fix(value): drop st_mode cast to satisfy MSRV clippy
CI's MSRV job (1.87, -D warnings) flagged `st.st_mode as u32` as
clippy::unnecessary_cast — st_mode is already mode_t (u32) on Linux. Newer
local clippy (1.95) didn't flag it, masking the failure. Use the field
directly; `as u32` and `u32::from(_)` each trip a lint when the source is
already u32, so the only clean form is no cast at all.

+++ src/value/fdinfo.rs
@@ -107,8 +107,8 @@ fn fstat_kind_size(raw: i32) -> (String, Option<u64>) {
     if unsafe { libc::fstat(raw, &mut st) } != 0 {
         return ("unknown".into(), None);
     }
-    // `st_mode` and the S_IF* constants are both mode_t (u32 on Linux).
-    let mode = st.st_mode as u32;
+    // `st_mode` is mode_t (u32 on Linux); the S_IF* constants are too.
+    let mode = st.st_mode;
     let fmt = mode & libc::S_IFMT;
     let kind = match fmt {
         libc::S_IFREG => "regular",
